Fix: correctly rewind tape
According to the specification, if M < 0, the tape is skipped backwards M
blocks, or to the beginning of the tape, whichever comes first.  In the
implementation, we don't check to verify that we aren't seeking past the
beginning of the file. This causes fseek(3) to fail, and it leaves us at the
position we were at.

Diagnosis and fix by Kevin Brunelle.